# scripts/generated-swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cooncierge Admin API",
    "version": "1.0.0",
    "description": "Comprehensive API documentation for Cooncierge Admin Server",
    "contact": {
      "name": "API Support",
      "email": "mfa@cooncierge.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    },
    {
      "url": "https://api.karvaann.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login/verify-2fa endpoints"
      },
      "karvaanToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-access-token",
        "description": "Karvaan token for authenticated requests"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "An error occurred"
          },
          "error": {
            "type": "string",
            "description": "Detailed error message"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/customer/get-all-customers": {
      "get": {
        "summary": "Get all customers",
        "description": "Retrieve all customers with populated owner information.\n\n**Features:**\n- Business-based filtering (users see only their business customers)\n- Soft deletion filtering\n- Owner information population\n- Multi-tenant architecture support\n",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "karvaanToken": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "isDeleted",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Filter by deletion status (true for deleted, false for active)"
          }
        ],
        "responses": {
          "200": {
            "description": "Customers retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "customers": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Customer"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch customers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/customer/get-customer/{id}": {
      "get": {
        "summary": "Get customer by ID",
        "description": "Retrieve a specific customer by ID with populated owner information",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "karvaanToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Customer ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "customer": {
                      "$ref": "#/components/schemas/Customer"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to get customer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/customer/create-customer": {
      "post": {
        "summary": "Create a new customer",
        "description": "Create a new customer record",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "karvaanToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "phone",
                  "ownerId"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Jane Smith"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "jane.smith@example.com"
                  },
                  "phone": {
                    "type": "string",
                    "example": "+1234567890"
                  },
                  "address": {
                    "type": "string",
                    "example": "123 Main St, City, State"
                  },
                  "ownerId": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439014"
                  },
                  "tier": {
                    "type": "string",
                    "enum": [
                      "tier1",
                      "tier2",
                      "tier3"
                    ],
                    "example": "tier1"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "customer": {
                      "$ref": "#/components/schemas/Customer"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to create customer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/customer/update-customer/{id}": {
      "put": {
        "summary": "Update a customer",
        "description": "Update an existing customer by ID",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "karvaanToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Customer ID to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "ownerId": {
                    "type": "string"
                  },
                  "tier": {
                    "type": "string",
                    "enum": [
                      "tier1",
                      "tier2",
                      "tier3"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "customer": {
                      "$ref": "#/components/schemas/Customer"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to update customer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/customer/delete-customer/{id}": {
      "delete": {
        "summary": "Delete a customer",
        "description": "Delete a customer by ID",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "karvaanToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Customer ID to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Customer deleted"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to delete customer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/customer/bulk-upload-template/{format}": {
      "get": {
        "summary": "Download bulk upload template file",
        "description": "Download a template file for bulk customer upload in the specified format.\nThe template includes sample data and all available columns with proper formatting.\n\n**Supported formats:**\n- csv: Comma-separated values format\n- xlsx: Microsoft Excel format\n\n**Template includes:**\n- All required columns: name, email, phone, ownerId\n- All optional columns: alias, dateOfBirth, gstin, companyName, openingBalance, balanceType, address, tier\n- Sample data rows showing proper formatting\n- Column headers for easy identification\n\n**Usage:**\n1. Download the template in your preferred format\n2. Fill in your customer data following the sample format\n3. Upload the completed file using POST /customer/bulk-upload\n\n**Important Notes:**\n- Replace sample ownerId values with actual team member IDs from your business\n- Ensure email addresses are unique within your business\n- Use proper date format (YYYY-MM-DD) for dateOfBirth\n- Use valid tier values: tier1, tier2, tier3, tier4, tier5\n- Use valid balanceType values: credit, debit\n",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "karvaanToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "format",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "csv",
                "xlsx"
              ]
            },
            "description": "Template file format (csv or xlsx)",
            "example": "csv"
          }
        ],
        "responses": {
          "200": {
            "description": "Template file downloaded successfully",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "description": "CSV template file with sample customer data"
                }
              },
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "description": "Excel template file with sample customer data"
                }
              }
            },
            "headers": {
              "Content-Disposition": {
                "description": "Attachment filename",
                "schema": {
                  "type": "string",
                  "example": "attachment; filename=\"customer-bulk-upload-template.csv\""
                }
              },
              "Content-Type": {
                "description": "File MIME type",
                "schema": {
                  "type": "string",
                  "examples": {
                    "csv": {
                      "value": "text/csv"
                    },
                    "xlsx": {
                      "value": "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid format parameter",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid format. Supported formats are: csv, xlsx"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error during template download"
                    },
                    "error": {
                      "type": "string",
                      "description": "Error details"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}